+++ internal/utils.go
@@ -3,6 +3,7 @@ package internal
 import (
 	"fmt"
 	"github.com/chalk-ai/chalk-go/internal/colls"
+	"github.com/chalk-ai/chalk-go/internal/ptr"
 	"github.com/cockroachdb/errors"
 	"math"
 	"os"
@@ -548,4 +549,18 @@ func BuildNamespaceMemo(memo NamespaceMemo, typ reflect.Type) error {
 		return BuildNamespaceMemo(memo, typ.Elem())
 	}
 	return nil
+}
+
+func getForeignNamespace(typ reflect.Type) *string {
+	if typ.Kind() == reflect.Ptr {
+		typ = typ.Elem()
+	}
+
+	if typ.Kind() == reflect.Slice {
+		return getForeignNamespace(typ.Elem())
+	} else if IsFeaturesClass(typ) {
+		return ptr.Ptr(ChalkpySnakeCase(typ.Name()))
+	} else {
+		return nil
+	}
 }
